+++ src/preamble.js
@@ -237,7 +237,6 @@ function postRun() {
 
 /**
  * @param {string|number=} what
- * @noreturn
  */
 function abort(what) {
 #if expectToReceiveOnModule('onAbort')