+++ test/codesize/test_codesize_minimal_O0.expected.js
@@ -510,7 +510,6 @@ function postRun() {
 
 /**
  * @param {string|number=} what
- * @noreturn
  */
 function abort(what) {