+++ src/lib/libwasm_worker.js
@@ -65,10 +65,10 @@ addToLibrary({
 #if PTHREADS
   // When the build contains both pthreads and Wasm Workers, offset the
   // Wasm Worker ID space to avoid collisions with pthread TIDs (which start
-  // at 42). We use `1 << 30` since it's ~1/2 way through `pid_t` space,
+  // at 42). We use `1 << 21` since it's ~1/2 way through `pid_t` space,
   // essentially giving pthreads the first 1/2 of the range and wasm workers the
   // second half.
-  $_wasmWorkersID: {{{ 1 << 30 }}},
+  $_wasmWorkersID: {{{ 1 << 21 }}},
 #else
   $_wasmWorkersID: 1,
 #endif
@@ -223,7 +223,17 @@ if (ENVIRONMENT_IS_WASM_WORKER
 #if ENVIRONMENT_MAY_BE_NODE
     if (ENVIRONMENT_IS_NODE) {
       /** @suppress {checkTypes} */
-      worker.on('message', (msg) => worker.onmessage({ data: msg }));
+      worker.on('message', (msg) => {
+        if (msg['cmd'] == 'uncaughtException') {
+          // Message handler for Node.js specific out-of-order behavior:
+          // https://github.com/nodejs/node/issues/59617
+          // A worker sent an uncaught exception event. Re-raise it on the main thread.
+          err(`worker sent an error! ${msg.error.message}`);
+          throw msg.error;
+        } else {
+          worker.onmessage({ data: msg });
+        }
+      });
     }
 #endif
 #if RUNTIME_DEBUG